requirements.txt
streamlit
langchain
langchain-core
langchain-ollama
langchain-milvus
langchain-text-splitters
langchain-community
pymupdf
python-docx
python-pptx
openpyxl
pymilvus
ollama
rank-bm25
flashrank
